Create separate tests for different packages
