+++ node/pkg/governor/governor_test.go
@@ -684,97 +684,3 @@ func TestIfAnythingIsQueuedEverythingIsQueued(t *testing.T) {
 	assert.Equal(t, 2, numPending)
 	assert.Equal(t, uint64(1774619+1774), valuePending)
 }
-
-func TestSerializeAndDeserializeOfTransfer(t *testing.T) {
-
-	tokenAddr, err := vaa.StringToAddress("0x707f9118e33a9b8998bea41dd0d46f38bb963fc8")
-	assert.Nil(t, err)
-
-	tokenBridgeAddr, _ := vaa.StringToAddress("0x0290fb167208af455bb137780163b7b7a9a10c16")
-	assert.Nil(t, err)
-
-	xfer1 := &db.Transfer{
-		Timestamp:      time.Unix(int64(1654516425), 0),
-		Value:          125000,
-		OriginChain:    vaa.ChainIDEthereum,
-		OriginAddress:  tokenAddr,
-		EmitterChain:   vaa.ChainIDEthereum,
-		EmitterAddress: tokenBridgeAddr,
-		MsgID:          "2/0000000000000000000000000290fb167208af455bb137780163b7b7a9a10c16/789101112131415",
-	}
-
-	bytes, err := xfer1.Marshal()
-	assert.Nil(t, err)
-
-	xfer2, err := db.UnmarshalTransfer(bytes)
-	assert.Nil(t, err)
-
-	assert.Equal(t, xfer1.Timestamp, xfer2.Timestamp)
-	assert.Equal(t, xfer1.Value, xfer2.Value)
-	assert.Equal(t, xfer1.OriginChain, xfer2.OriginChain)
-	assert.Equal(t, xfer1.OriginAddress, xfer2.OriginAddress)
-	assert.Equal(t, xfer1.EmitterChain, xfer2.EmitterChain)
-	assert.Equal(t, xfer1.EmitterAddress, xfer2.EmitterAddress)
-	assert.Equal(t, xfer1.MsgID, xfer2.MsgID)
-
-	expectedTransferKey := "GOV:XFER:2/0000000000000000000000000290fb167208af455bb137780163b7b7a9a10c16/789101112131415"
-	assert.Equal(t, expectedTransferKey, string(db.TransferMsgID(xfer2)))
-}
-
-func TestSerializeAndDeserializeOfMessagePublication(t *testing.T) {
-	tokenAddrStr := "0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"
-	toAddrStr := "0x707f9118e33a9b8998bea41dd0d46f38bb963fc8"
-	tokenBridgeAddr, _ := vaa.StringToAddress("0x0290fb167208af455bb137780163b7b7a9a10c16")
-
-	payloadBytes1 := buildMockTransferPayloadBytes(1,
-		vaa.ChainIDEthereum,
-		tokenAddrStr,
-		vaa.ChainIDPolygon,
-		toAddrStr,
-		270,
-	)
-
-	msg1 := &common.MessagePublication{
-		TxHash:           eth_common.HexToHash("0x06f541f5ecfc43407c31587aa6ac3a689e8960f36dc23c332db5510dfc6a4063"),
-		Timestamp:        time.Unix(int64(1654516425), 0),
-		Nonce:            123456,
-		Sequence:         789101112131415,
-		EmitterChain:     vaa.ChainIDEthereum,
-		EmitterAddress:   tokenBridgeAddr,
-		Payload:          payloadBytes1,
-		ConsistencyLevel: 16,
-	}
-
-	bytes, err := msg1.Marshal()
-	assert.Nil(t, err)
-
-	msg2, err := common.UnmarshalMessagePublication(bytes)
-	assert.Nil(t, err)
-
-	assert.Equal(t, msg1.TxHash, msg2.TxHash)
-	assert.Equal(t, msg1.Timestamp, msg2.Timestamp)
-	assert.Equal(t, msg1.Nonce, msg2.Nonce)
-	assert.Equal(t, msg1.Sequence, msg2.Sequence)
-	assert.Equal(t, msg1.EmitterChain, msg2.EmitterChain)
-	assert.Equal(t, msg1.EmitterAddress, msg2.EmitterAddress)
-	assert.Equal(t, msg1.ConsistencyLevel, msg2.ConsistencyLevel)
-
-	payload2, err := vaa.DecodeTransferPayloadHdr(msg2.Payload)
-	assert.Nil(t, err)
-
-	expectTokenAddr, err := vaa.StringToAddress(tokenAddrStr)
-	assert.Nil(t, err)
-
-	expectToAddr, err := vaa.StringToAddress(toAddrStr)
-	assert.Nil(t, err)
-
-	assert.Equal(t, uint8(1), payload2.Type)
-	assert.Equal(t, vaa.ChainIDEthereum, payload2.OriginChain)
-	assert.Equal(t, expectTokenAddr, payload2.OriginAddress)
-	assert.Equal(t, vaa.ChainIDPolygon, payload2.TargetChain)
-	assert.Equal(t, expectToAddr, payload2.TargetAddress)
-	assert.Equal(t, 0, big.NewInt(27000000000).Cmp(payload2.Amount))
-
-	expectedPendingKey := "GOV:PENDING:2/0000000000000000000000000290fb167208af455bb137780163b7b7a9a10c16/789101112131415"
-	assert.Equal(t, expectedPendingKey, string(db.PendingMsgID(msg2)))
-}
